06.09.2022
- Player Movement independent from deltaTime
- Improved Screenflash system | Screenflash for Electrify, Heal sound for Medkit
- Timer ticking sound for Flip running out
- Fixed Shadowtraces not giving speed
- Fixed Leech dealing 0.4 dmg | Classic Mode phase dmg

+++ SSS222/Assets/Scripts/Core/SaveSerial.cs
@@ -318,9 +318,9 @@ public class SaveSerial : MonoBehaviour{	public static SaveSerial instance;
 [System.Serializable]
 public class Highscore{
 	public int score;
-	public float playtime;
+	public int playtime;
 	public string version;
-	public float build;
+	public double build;
 	public DateTime date;
 }
 [System.Serializable]
